split monolithic tests into individual test files
