+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/server/execution/CustomExecutorsTest.java
@@ -104,20 +104,20 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 
-@RunWith(MockitoJUnitRunner.class)
-public class CustomExecutorsTest {
+public abstract class CustomExecutorsTest {
 
-  private static final Schema SCHEMA =
+  static final Schema SCHEMA =
       SchemaBuilder.struct().field("val", Schema.OPTIONAL_STRING_SCHEMA).build();
 
   @Rule
   public ExpectedException expectedException = ExpectedException.none();
 
-  private KsqlConfig ksqlConfig;
   private KsqlEngine realEngine;
-  private KsqlEngine engine;
   private MutableMetaStore metaStore;
-  private ServiceContext serviceContext;
+
+  KsqlConfig ksqlConfig;
+  KsqlEngine engine;
+  ServiceContext serviceContext;
 
   @Before
   public void setUp() {
@@ -142,542 +142,9 @@ public class CustomExecutorsTest {
     serviceContext.close();
   }
 
-  @Test
-  public void shouldListKafkaTopics() {
-    // Given:
-    givenKsqlTopic("topic1");
-    givenKafkaTopic("topic2");
-
-    final AdminClient mockAdminClient = mock(AdminClient.class);
-    final ListConsumerGroupsResult result = mock(ListConsumerGroupsResult.class);
-    final KafkaFutureImpl<Collection<ConsumerGroupListing>> groups = new KafkaFutureImpl<>();
-
-    when(result.all()).thenReturn(groups);
-    when(mockAdminClient.listConsumerGroups()).thenReturn(result);
-    groups.complete(ImmutableList.of());
-
-    final ServiceContext serviceContext = TestServiceContext.create(
-        this.serviceContext.getKafkaClientSupplier(),
-        mockAdminClient,
-        this.serviceContext.getTopicClient(),
-        this.serviceContext.getSchemaRegistryClientFactory()
-    );
-
-     // When:
-    final KafkaTopicsList topicsList =
-        (KafkaTopicsList) CustomExecutors.LIST_TOPICS.execute(
-        prepare("LIST TOPICS;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(topicsList.getTopics(), containsInAnyOrder(
-        new KafkaTopicInfo("topic1", true, ImmutableList.of(1), 0, 0),
-        new KafkaTopicInfo("topic2", false, ImmutableList.of(1), 0, 0)
-    ));
-  }
-
-  @Test
-  public void shouldListRegisteredTopics() {
-    // Given:
-    final KsqlTopic topic1 = givenKsqlTopic("topic1");
-    final KsqlTopic topic2 = givenKsqlTopic("topic2");
-
-    // When:
-    final KsqlTopicsList topicsList =
-        (KsqlTopicsList) CustomExecutors.LIST_REGISTERED_TOPICS.execute(
-        prepare("LIST REGISTERED TOPICS;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(topicsList.getTopics(),
-        contains(new KsqlTopicInfo(topic1), new KsqlTopicInfo(topic2)));
-  }
-
-  @Test
-  public void shouldListQueriesEmpty() {
-    // When
-    final Queries queries = (Queries) CustomExecutors.LIST_QUERIES.execute(
-        prepare("SHOW QUERIES;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    assertThat(queries.getQueries(), is(empty()));
-  }
-
-  @Test
-  public void shouldListQueriesBasic() {
-    // Given
-    final PreparedStatement showQueries = prepare("SHOW QUERIES;");
-    final PersistentQueryMetadata metadata = givenPersistentQuery("id");
-
-    engine = mock(KsqlEngine.class);
-    when(engine.getPersistentQueries()).thenReturn(ImmutableList.of(metadata));
-
-    // When
-    final Queries queries = (Queries) CustomExecutors.LIST_QUERIES.execute(
-        showQueries,
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    assertThat(queries.getQueries(), containsInAnyOrder(
-        new RunningQuery(
-            metadata.getStatementString(),
-            metadata.getSinkNames(),
-            new EntityQueryId(metadata.getQueryId()))));
-  }
-
-  @Test
-  public void shouldListQueriesExtended() {
-    // Given
-    final PreparedStatement showQueries = prepare("SHOW QUERIES EXTENDED;");
-    final PersistentQueryMetadata metadata = givenPersistentQuery("id");
-
-    engine = mock(KsqlEngine.class);
-    when(engine.getPersistentQueries()).thenReturn(ImmutableList.of(metadata));
-
-    // When
-    final QueryDescriptionList queries = (QueryDescriptionList) CustomExecutors.LIST_QUERIES.execute(
-        showQueries,
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    assertThat(queries.getQueryDescriptions(), containsInAnyOrder(
-        QueryDescription.forQueryMetadata(metadata)));
-  }
-
-  @Test
-  public void shouldListFunctions() {
-    // When:
-    final FunctionNameList functionList = (FunctionNameList) CustomExecutors.LIST_FUNCTIONS.execute(
-        prepare("LIST FUNCTIONS;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(functionList.getFunctions(), hasItems(
-        new SimpleFunctionInfo("EXTRACTJSONFIELD", FunctionType.scalar),
-        new SimpleFunctionInfo("ARRAYCONTAINS", FunctionType.scalar),
-        new SimpleFunctionInfo("CONCAT", FunctionType.scalar),
-        new SimpleFunctionInfo("TOPK", FunctionType.aggregate),
-        new SimpleFunctionInfo("MAX", FunctionType.aggregate)));
-
-    assertThat("shouldn't contain internal functions", functionList.getFunctions(),
-        not(hasItem(new SimpleFunctionInfo("FETCH_FIELD_FROM_STRUCT", FunctionType.scalar))));
-  }
-
-  @Test
-  public void shouldListProperties() {
-    // When:
-    final PropertiesList properties = (PropertiesList) CustomExecutors.LIST_PROPERTIES.execute(
-        prepare("LIST PROPERTIES;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(properties.getProperties(),
-        equalTo(ksqlConfig.getAllConfigPropsWithSecretsObfuscated()));
-    assertThat(properties.getOverwrittenProperties(), is(empty()));
-  }
-
-  @Test
-  public void shouldListPropertiesWithOverrides() {
-    // When:
-    final PropertiesList properties = (PropertiesList) CustomExecutors.LIST_PROPERTIES.execute(
-        prepare("LIST PROPERTIES;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of("auto.offset.reset", "latest")
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(properties.getProperties(),
-        hasEntry("ksql.streams.auto.offset.reset", "latest"));
-    assertThat(properties.getOverwrittenProperties(), hasItem("ksql.streams.auto.offset.reset"));
-  }
-
-  @Test
-  public void shouldNotListSslProperties() {
-    // When:
-    final PropertiesList properties = (PropertiesList) CustomExecutors.LIST_PROPERTIES.execute(
-        prepare("LIST PROPERTIES;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(properties.getProperties(), not(hasKey(isIn(KsqlConfig.SSL_CONFIG_NAMES))));
-  }
-
-  @Test
-  public void shouldShowStreams() {
-    // Given:
-    final KsqlStream stream1 = givenSource(DataSourceType.KSTREAM, "stream1");
-    final KsqlStream stream2 = givenSource(DataSourceType.KSTREAM, "stream2");
-    givenSource(DataSourceType.KTABLE, "table");
-
-    // When:
-    final StreamsList descriptionList = (StreamsList)
-        CustomExecutors.LIST_STREAMS.execute(
-            prepare("SHOW STREAMS;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(descriptionList.getStreams(), containsInAnyOrder(
-        new SourceInfo.Stream(stream1),
-        new SourceInfo.Stream(stream2)
-    ));
-  }
-
-  @Test
-  public void shouldShowStreamsExtended() {
-    // Given:
-    final KsqlStream stream1 = givenSource(DataSourceType.KSTREAM, "stream1");
-    final KsqlStream stream2 = givenSource(DataSourceType.KSTREAM, "stream2");
-    givenSource(DataSourceType.KTABLE, "table");
-
-    // When:
-    final SourceDescriptionList descriptionList = (SourceDescriptionList)
-        CustomExecutors.LIST_STREAMS.execute(
-            prepare("SHOW STREAMS EXTENDED;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-        ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(descriptionList.getSourceDescriptions(), containsInAnyOrder(
-        new SourceDescription(stream1, true, "JSON", ImmutableList.of(), ImmutableList.of(), null),
-        new SourceDescription(stream2, true, "JSON", ImmutableList.of(), ImmutableList.of(), null)
-    ));
-  }
-
-  @Test
-  public void shouldShowTables() {
-    // Given:
-    final KsqlTable table1 = givenSource(DataSourceType.KTABLE, "table1");
-    final KsqlTable table2 = givenSource(DataSourceType.KTABLE, "table2");
-    givenSource(DataSourceType.KSTREAM, "stream");
-
-    // When:
-    final TablesList descriptionList = (TablesList)
-        CustomExecutors.LIST_TABLES.execute(
-            prepare("LIST TABLES;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-        ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(descriptionList.getTables(), containsInAnyOrder(
-        new SourceInfo.Table(table1),
-        new SourceInfo.Table(table2)
-    ));
-  }
-
-  @Test
-  public void shouldShowTablesExtended() {
-    // Given:
-    final KsqlTable table1 = givenSource(DataSourceType.KTABLE, "table1");
-    final KsqlTable table2 = givenSource(DataSourceType.KTABLE, "table2");
-    givenSource(DataSourceType.KSTREAM, "stream");
-
-    // When:
-    final SourceDescriptionList descriptionList = (SourceDescriptionList)
-        CustomExecutors.LIST_TABLES.execute(
-            prepare("LIST TABLES EXTENDED;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-        ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    final KafkaTopicClient client = serviceContext.getTopicClient();
-    assertThat(descriptionList.getSourceDescriptions(), containsInAnyOrder(
-        new SourceDescription(table1, true, "JSON", ImmutableList.of(), ImmutableList.of(), client),
-        new SourceDescription(table2, true, "JSON", ImmutableList.of(), ImmutableList.of(), client)
-    ));
-  }
-
-  @Test
-  public void shouldShowColumnsSource() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "SOURCE");
-    final ExecuteResult result = engine.execute(
-        prepare("CREATE STREAM SINK AS SELECT * FROM source;"),
-        ksqlConfig,
-        ImmutableMap.of());
-    final PersistentQueryMetadata metadata = (PersistentQueryMetadata) result.getQuery()
-        .orElseThrow(IllegalArgumentException::new);
-    final StructuredDataSource stream = engine.getMetaStore().getSource("SINK");
-
-    // When:
-    final SourceDescriptionEntity sourceDescription = (SourceDescriptionEntity)
-        CustomExecutors.SHOW_COLUMNS.execute(
-            PreparedStatement.of(
-                "DESCRIBE SINK;",
-                new ShowColumns(QualifiedName.of("SINK"), false, false)),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(sourceDescription.getSourceDescription(),
-        equalTo(new SourceDescription(
-            stream,
-            false,
-            "JSON",
-            ImmutableList.of(),
-            ImmutableList.of(new RunningQuery(
-                metadata.getStatementString(),
-                metadata.getSinkNames(),
-                new EntityQueryId(metadata.getQueryId()))),
-            null)));
-  }
-
-  @Test
-  public void shouldShowColumnsTopic() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "S");
-
-    // When:
-    final TopicDescription description = (TopicDescription) CustomExecutors.SHOW_COLUMNS.execute(
-        prepare("DESCRIBE TOPIC S;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(description,
-        equalTo(new TopicDescription("DESCRIBE TOPIC S;", "S", "S", "JSON", SCHEMA.toString())));
-  }
-
-  @Test
-  public void shouldThrowOnDescribeMissingTopic() {
-    // Expect:
-    expectedException.expect(KsqlException.class);
-    expectedException.expectMessage("Could not find Topic");
-
-    // When:
-    CustomExecutors.SHOW_COLUMNS.execute(
-        prepare("DESCRIBE TOPIC S;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    );
-  }
-
-  @Test
-  public void shouldSetProperty() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "stream");
-    final Map<String, Object> properties = new HashMap<>();
-
-    // When:
-    CustomExecutors.SET_PROPERTY.execute(
-        prepare("SET 'property' = 'value';"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        properties
-    );
-
-    // Then:
-    assertThat(properties, hasEntry("property", "value"));
-  }
-
-  @Test
-  public void shouldUnSetProperty() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "stream");
-    final Map<String, Object> properties = new HashMap<>();
-    properties.put("property", "value");
-
-    // When:
-    CustomExecutors.UNSET_PROPERTY.execute(
-        prepare("UNSET 'property';"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        properties
-    );
-
-    // Then:
-    assertThat(properties, not(hasKey("property")));
-  }
-
-  @Test
-  public void shouldDescribeUDF() {
-    // When:
-    final FunctionDescriptionList functionList = (FunctionDescriptionList)
-        CustomExecutors.DESCRIBE_FUNCTION.execute(
-            prepare("DESCRIBE FUNCTION CONCAT;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-        ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(functionList, new TypeSafeMatcher<FunctionDescriptionList>() {
-      @Override
-      protected boolean matchesSafely(FunctionDescriptionList item) {
-        return functionList.getName().equals("CONCAT")
-            && functionList.getType().equals(FunctionType.scalar);
-      }
-
-      @Override
-      public void describeTo(Description description) {
-        description.appendText(functionList.getName());
-      }
-    });
-  }
-
-  @Test
-  public void shouldDescribeUDAF() {
-    // When:
-    final FunctionDescriptionList functionList = (FunctionDescriptionList)
-        CustomExecutors.DESCRIBE_FUNCTION.execute(
-            prepare("DESCRIBE FUNCTION MAX;"),
-            engine,
-            serviceContext,
-            ksqlConfig,
-            ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(functionList, new TypeSafeMatcher<FunctionDescriptionList>() {
-      @Override
-      protected boolean matchesSafely(FunctionDescriptionList item) {
-        return functionList.getName().equals("MAX")
-            && functionList.getType().equals(FunctionType.aggregate);
-      }
-
-      @Override
-      public void describeTo(Description description) {
-        description.appendText(functionList.getName());
-      }
-    });
-  }
-
-  @Test
-  public void shouldExplainQueryId() {
-    // Given:
-    final PreparedStatement explain = prepare("EXPLAIN id;");
-    final PersistentQueryMetadata metadata = givenPersistentQuery("id");
-
-    engine = mock(KsqlEngine.class);
-    when(engine.getPersistentQuery(metadata.getQueryId())).thenReturn(Optional.of(metadata));
-
-    // When:
-    final QueryDescriptionEntity query = (QueryDescriptionEntity) CustomExecutors.EXPLAIN.execute(
-        explain,
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(query.getQueryDescription(), equalTo(QueryDescription.forQueryMetadata(metadata)));
-  }
-
-  @Test
-  public void shouldExplainPersistentStatement() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "Y");
-    final String statementText = "CREATE STREAM X AS SELECT * FROM Y;";
-    final PreparedStatement explain = prepare("EXPLAIN " + statementText);
-
-    // When:
-    final QueryDescriptionEntity query = (QueryDescriptionEntity) CustomExecutors.EXPLAIN.execute(
-        explain,
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(query.getQueryDescription().getStatementText(), equalTo(statementText));
-    assertThat(query.getQueryDescription().getSources(), containsInAnyOrder("Y"));
-    assertThat("No side effects should happen", engine.getPersistentQueries(), is(empty()));
-  }
-
-  @Test
-  public void shouldExplainStatement() {
-    // Given:
-    givenSource(DataSourceType.KSTREAM, "Y");
-    final String statementText = "SELECT * FROM Y;";
-    final PreparedStatement explain = prepare("EXPLAIN " + statementText);
-
-    // When:
-    final QueryDescriptionEntity query = (QueryDescriptionEntity) CustomExecutors.EXPLAIN.execute(
-        explain,
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    ).orElseThrow(IllegalStateException::new);
-
-    // Then:
-    assertThat(query.getQueryDescription().getStatementText(), equalTo(statementText));
-    assertThat(query.getQueryDescription().getSources(), containsInAnyOrder("Y"));
-  }
-
-  @Test
-  public void shouldFailOnNonQueryExplain() {
-    // Expect:
-    expectedException.expect(KsqlException.class);
-    expectedException.expectMessage("The provided statement does not run a ksql query");
-
-    // When:
-    CustomExecutors.EXPLAIN.execute(
-        prepare("Explain SHOW TOPICS;"),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    );
-  }
 
   @SuppressWarnings("unchecked")
-  private <T extends StructuredDataSource> T givenSource(
+  <T extends StructuredDataSource> T givenSource(
       final DataSource.DataSourceType type,
       final String name) {
     final KsqlTopic topic = givenKsqlTopic(name);
@@ -706,24 +173,24 @@ public class CustomExecutorsTest {
     return (T) source;
   }
 
-  private KsqlTopic givenKsqlTopic(String name) {
+  KsqlTopic givenKsqlTopic(String name) {
     final KsqlTopic topic = new KsqlTopic(name, name, new KsqlJsonTopicSerDe(), false);
     givenKafkaTopic(name);
     metaStore.putTopic(topic);
     return topic;
   }
 
-  private void givenKafkaTopic(final String name) {
+  void givenKafkaTopic(final String name) {
     ((FakeKafkaTopicClient) serviceContext.getTopicClient())
         .preconditionTopicExists(name, 1, (short) 1, Collections.emptyMap());
   }
 
-  private PreparedStatement prepare(final String sql) {
+  PreparedStatement prepare(final String sql) {
     return engine.prepare(new DefaultKsqlParser().parse(sql).get(0));
   }
 
   @SuppressWarnings("SameParameterValue")
-  private PersistentQueryMetadata givenPersistentQuery(final String id) {
+  PersistentQueryMetadata givenPersistentQuery(final String id) {
     final PersistentQueryMetadata metadata = mock(PersistentQueryMetadata.class);
     when(metadata.getQueryId()).thenReturn(new QueryId(id));
     when(metadata.getSinkNames()).thenReturn(ImmutableSet.of(id));
