+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/server/execution/PropertyExecutorTest.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright 2019 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License; you may not use this file
+ * except in compliance with the License.  You may obtain a copy of the License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.rest.server.execution;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.hamcrest.Matchers.hasKey;
+import static org.hamcrest.Matchers.not;
+
+import io.confluent.ksql.serde.DataSource.DataSourceType;
+import java.util.HashMap;
+import java.util.Map;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class PropertyExecutorTest extends CustomExecutorsTest {
+
+  @Test
+  public void shouldSetProperty() {
+    // Given:
+    givenSource(DataSourceType.KSTREAM, "stream");
+    final Map<String, Object> properties = new HashMap<>();
+
+    // When:
+    CustomExecutors.SET_PROPERTY.execute(
+        prepare("SET 'property' = 'value';"),
+        engine,
+        serviceContext,
+        ksqlConfig,
+        properties
+    );
+
+    // Then:
+    assertThat(properties, hasEntry("property", "value"));
+  }
+
+  @Test
+  public void shouldUnSetProperty() {
+    // Given:
+    givenSource(DataSourceType.KSTREAM, "stream");
+    final Map<String, Object> properties = new HashMap<>();
+    properties.put("property", "value");
+
+    // When:
+    CustomExecutors.UNSET_PROPERTY.execute(
+        prepare("UNSET 'property';"),
+        engine,
+        serviceContext,
+        ksqlConfig,
+        properties
+    );
+
+    // Then:
+    assertThat(properties, not(hasKey("property")));
+  }
+
+
+}
