+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/server/validation/CustomValidatorsTest.java
@@ -50,23 +50,22 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 
-@RunWith(MockitoJUnitRunner.class)
 public class CustomValidatorsTest {
 
   @Rule
   public ExpectedException expectedException = ExpectedException.none();
 
-  private KsqlConfig ksqlConfig;
+  KsqlConfig ksqlConfig;
+  ServiceContext serviceContext;
+  KsqlEngine engine;
+
   private KsqlEngine realEngine;
-  private ServiceContext serviceContext;
-  private KsqlEngine engine;
-  private MutableMetaStore metaStore;
 
   @Before
   public void setUp() {
     ksqlConfig = new KsqlConfig(new HashMap<>());
     serviceContext = TestServiceContext.create();
-    metaStore = new MetaStoreImpl(new InternalFunctionRegistry());
+    MutableMetaStore metaStore = new MetaStoreImpl(new InternalFunctionRegistry());
     engine = KsqlEngineTestUtil.createKsqlEngine(serviceContext, metaStore);
     realEngine = engine;
   }
@@ -77,140 +76,8 @@ public class CustomValidatorsTest {
     serviceContext.close();
   }
 
-  @Test
-  public void shouldThrowExceptionOnQueryEndpoint() {
-    // Expect:
-    expectBadRequest(
-        "SELECT and PRINT queries must use the /query endpoint",
-        "SELECT * FROM test_table;");
-
-    // When:
-    CustomValidators.QUERY_ENDPOINT.validate(
-        PreparedStatement.of("SELECT * FROM test_table;", mock(Query.class)),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    );
-  }
-
-  @Test
-  public void shouldThrowExceptionOnPrintTopic() {
-    // Expect:
-    expectBadRequest(
-        "SELECT and PRINT queries must use the /query endpoint",
-        "PRINT 'topic';");
-
-    // When:
-    CustomValidators.PRINT_TOPIC.validate(
-        PreparedStatement.of("PRINT 'topic';", mock(PrintTopic.class)),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    );
-  }
-
-  @Test
-  public void shouldFailOnUnknownSetProperty() {
-    // Expect:
-    expectedException.expect(KsqlStatementException.class);
-    expectedException.expectMessage("Unknown property: consumer.invalid");
-
-    // When:
-    CustomValidators.SET_PROPERTY.validate(
-        PreparedStatement.of(
-            "SET 'consumer.invalid'='value';",
-            new SetProperty(Optional.empty(), "consumer.invalid", "value")),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        new HashMap<>()
-    );
-  }
-
-  @Test
-  public void shouldAllowSetKnownProperty() {
-    // No exception when:
-    CustomValidators.SET_PROPERTY.validate(
-        PreparedStatement.of(
-            "SET '" + KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY + "' = '1';",
-            new SetProperty(Optional.empty(), KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY, "1")),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        new HashMap<>()
-    );
-  }
-
-  @Test
-  public void shouldFailOnInvalidSetPropertyValue() {
-    // Expect:
-    expectedException.expect(KsqlStatementException.class);
-    expectedException.expectMessage("Invalid value invalid");
-
-    // When:
-    CustomValidators.SET_PROPERTY.validate(
-        PreparedStatement.of(
-            "SET '" + KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY + "' = 'invalid';",
-            new SetProperty(Optional.empty(), KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY, "invalid")),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        new HashMap<>()
-    );
-  }
-
-  @Test
-  public void shouldFailOnUnknownUnsetProperty() {
-    // Expect:
-    expectedException.expect(KsqlStatementException.class);
-    expectedException.expectMessage("Unknown property: consumer.invalid");
-
-    // When:
-    CustomValidators.UNSET_PROPERTY.validate(
-        PreparedStatement.of(
-            "UNSET 'consumer.invalid';",
-            new UnsetProperty(Optional.empty(), "consumer.invalid")),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        new HashMap<>()
-    );
-  }
-
-  @Test
-  public void shouldAllowUnsetKnownProperty() {
-    // No exception when:
-    CustomValidators.UNSET_PROPERTY.validate(
-        PreparedStatement.of(
-            "UNSET '" + KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY + "';",
-            new UnsetProperty(Optional.empty(), KsqlConfig.SINK_NUMBER_OF_REPLICAS_PROPERTY)),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        new HashMap<>()
-    );
-  }
-
-  @Test
-  public void shouldFailOnTerminateUnknownQueryId() {
-    // Expect:
-    expectedException.expect(KsqlStatementException.class);
-    expectedException.expectMessage("Unknown queryId");
-
-    // When:
-    CustomValidators.TERMINATE_QUERY.validate(
-        PreparedStatement.of("", new TerminateQuery("id")),
-        engine,
-        serviceContext,
-        ksqlConfig,
-        ImmutableMap.of()
-    );
-  }
-
   @SuppressWarnings("SameParameterValue")
-  private void expectBadRequest(final String message, final String statementText) {
+  void expectBadRequest(final String message, final String statementText) {
     expectedException.expect(KsqlRestException.class);
     expectedException.expect(exceptionStatusCode(is(Code.BAD_REQUEST)));
     expectedException.expect(exceptionStatementErrorMessage(errorMessage(containsString(
